Renamed `interaction.slash` to `interaction.slash_command`
