--- src/main/java/com/dynxsty/dih4jda/commands/interactions/context/dao/GlobalContextCommand.java
@@ -1,4 +1,0 @@
-package com.dynxsty.dih4jda.commands.interactions.context.dao;
-
-public class GlobalContextCommand extends BaseContextCommand {
-}
