+++ src/main/java/com/dynxsty/dih4jda/commands/interactions/context_command/dao/GuildContextCommand.java
@@ -1,0 +1,4 @@
+package com.dynxsty.dih4jda.commands.interactions.context_command.dao;
+
+public class GuildContextCommand extends BaseContextCommand {
+}
