+++ src/main/java/com/dynxsty/dih4jda/commands/interactions/slash_command/dao/GlobalSlashCommand.java
@@ -1,3 +1,3 @@
-package com.dynxsty.dih4jda.commands.interactions.slash.dao;
+package com.dynxsty.dih4jda.commands.interactions.slash_command.dao;
 
 public abstract class GlobalSlashCommand extends BaseSlashCommand {}
