+++ src/main/java/com/dynxsty/dih4jda/commands/interactions/slash_command/dao/SlashSubcommandGroup.java
@@ -1,4 +1,4 @@
-package com.dynxsty.dih4jda.commands.interactions.slash.dao;
+package com.dynxsty.dih4jda.commands.interactions.slash_command.dao;
 
 import net.dv8tion.jda.api.interactions.commands.build.SubcommandGroupData;
